+++ History.md
@@ -2,6 +2,7 @@
 
 ### New Features
 * Ruby 2.0.0 support (#377 Matt Wynne & #357 @charliesome)
+* Capture duration value for json formatter (#329 Rick Beyer)
 * Added support for Hindi (hi), although some systems may need to install fonts which support the Devanagari script.
 * Obey program suffix when invoking bundler (#324 Eric Hodel)